+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/util/ChannelUtils.kt
@@ -3,16 +3,13 @@ package io.getstream.chat.android.compose.ui.util
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ReadOnlyComposable
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.buildAnnotatedString
 import com.getstream.sdk.chat.viewmodel.messages.getCreatedAtOrThrow
-import io.getstream.chat.android.client.extensions.getUsersExcludingCurrent
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
-import io.getstream.chat.android.compose.R
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 import io.getstream.chat.android.compose.viewmodel.channel.ChannelListViewModel
 import java.util.Date
@@ -29,21 +26,6 @@ public var Channel.isMuted: Boolean
     set(value) {
         extraData[EXTRA_CHANNEL_MUTED] = value
     }
-
-/**
- * Generates the display name for a channel based on its data.
- *
- * @return The display name of the channel.
- */
-@Composable
-@ReadOnlyComposable
-public fun Channel.getDisplayName(): String {
-    return name.takeIf { it.isNotEmpty() }
-        ?: getUsersExcludingCurrent()
-            .joinToString { it.name }
-            .takeIf { it.isNotEmpty() }
-        ?: stringResource(id = R.string.stream_compose_untitled_channel)
-}
 
 /**
  * Returns channel's last regular or system message if exists.