+++ CHANGELOG.md
@@ -82,6 +82,7 @@
 - Added the `emptySearchContent` parameter to `ChannelList` component that allows to customize the empty placeholder, when there are no channels matching the search query.
 - Fixed channel options that are displayed in the `ChannelInfo` component.
 - Added support for the muted channel indicator in the message list.
+- Added `ChannelNameFormatter` option to the `ChatTheme`, to allow for channel name format customization across the app.
 
 ### 🐞 Fixed